// workspaces/scorecard/plugins/scorecard/src/components/AggregatedMetricCards/AverageCard/types.ts
/*
 * Copyright Red Hat, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {
  AggregatedMetricAverageResult,
  AggregatedMetricResult,
} from '@red-hat-developer-hub/backstage-plugin-scorecard-common';

export type AverageCardComponentProps = {
  scorecard: Omit<AggregatedMetricResult, 'result'> & {
    result: AggregatedMetricAverageResult;
  };
  cardTitle: string;
  description: string;
  aggregationId: string;
  showSubheader?: boolean;
  showInfo?: boolean;
  dataTestId?: string;
};

export type TooltipPosition = { left: number; top: number };